requirements.txt
MetaTrader5
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0
pytz>=2023.0
python-dotenv>=1.0.0
